+++ src/options/components/Header/Header.tsx
@@ -1,6 +1,6 @@
 import { Wrapper } from './Header.styled';
 import { H1 } from '../Headings/Headings';
-import { i18n } from '../i18n/i18n';
+import { i18n } from '../../utils/i18n/i18n';
 
 export const Header = () => (
   <Wrapper>
