+++ src/options/components/HowTo/HowTo.tsx
@@ -1,5 +1,5 @@
 import { useState, useCallback, useMemo } from 'react';
-import { i18n } from '../i18n/i18n';
+import { i18n } from '../../utils/i18n/i18n';
 import { H2 } from '../Headings/Headings';
 import { LinkButton } from '../LinkButton/LinkButton';
 import { Modal } from '../Modal/Modal';
