+++ src/options/utils/i18n/i18n.ts
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react';
 import { getReplacementRegExp } from '../../../shared/getReplacementRegExp/getReplacementRegExp';
 import { getTranslatedText, Key } from '../../../shared/getTranslatedText/getTranslatedText';
-import { replaceWithElements } from '../utils/replaceWithElements';
+import { replaceWithElements } from '../../components/utils/replaceWithElements';
 
 export function i18n(children: Key): string;
 export function i18n(children: Key, substitutions: ReactNode[]): ReactNode;
